Make N-1 attack work
I found a compound in the "member" samples (samples that are used in training) which has a rare non-zero index. The compound is the 26822nd and the index is 2945th.

Rare non-zero indexes produce a ~0.3 non-zero rate in the corresponding gradient, thus the default 0.2 voting threshold. Common indexes produce ~0.5 non-zero rate in connected gradients consistently (regardless of whether the sample was in the batch or not).